hooked up assets, moved to hidden files, refactor
Assets now work as we advertised they would.
```
import fileUrl from "assets/images/test.jpg";
```

Moved config files that we intended to be hidden back to hidden files.
It seems as though the node module would ignore files with the name
`._something` but it didn't mind `.something` so we moved them. This
will help avoid people trying to figure out what those files are and it
will hopefully stop people from trying to edit them.

The assets folder and actions empty folders are now created in the
intialize new project step. We just had to include a file in the empty
folders so that they are not empty. They now include a file named
.empty.

Made an update to the `createStore` helper method so that it includes a
reducer that always returns the same state no matter what the action.
This prevents us from getting the error in the console that was
complaining there were no reducers.