Rename empty reducer to _gluestick, fix node 4 bug
emptyReducer was added so that there was at least 1 reducer for the
store to use. The name emptyReducer wasn't very useful and the data
wasn't valuable as the sub state was null. I have renamed the reducer to
_gluestick which will return the value of true. This still isn't
extremely useful but at least now you can use it to detect if a project
was generated using gluestick.

There was also a dependency issue when using node 4.x. One of our
dependencies `history` was not installing it's own dependencies. By
bumping the version of `history` this problem seems to have been
resolved.