Fix npm start
When you run `npm start` it will run `gluestick start` but it doesn't
run from /usr/local/bin it actually runs it from the project's
node_moduls/.bin/gluestick location. For this reason we cannot ignore
the `cli.js` file or we wont be able to use ES6 in our command line
interface.  do not skip the babel hook on the cli file