Fix bug with `webpack-additions` missing file
We use `get-webpack-additions` to get the data back because it will first
check if the file exists before trying to include it, falling back to empty
arrays for defaults. Requiring get-webpack-additions will return a function
instead of the result so we can execute the check asynchronously vs making
it part of the initial build. If it were a simple require it would throw an
error when babel or webpack try to resolve missing dependencies. We then
call this method with `true` as the first and only argument because that is
how we tell the method we are requiring for the purpose of the isomorphic
tools, not the webpack config file.