Add LazyMethodRequire method
This method is used when you want to import a method that shouldn't be
resolved until it is called. This prevents issues in commands like
`start-client` that rely on being run from within an existing project.
These commands would look in `process.cwd()` for files that don't exist when
you are running a command line `gluestick new …`. This way, imports only
happen when they are needed.