Add support to load css files from npm modules
Use case: `import 'bootstrap/dist/css/bootstrap.css';`

This was causing problems for 2 reasons.
1. The bootstrap.css file was references fonts outside of the `assets`
folder. Our loader was limited to only loading fonts from the `assets`
folder. There is now a good reason why we should not limit it to the
that folder and I couldn't think of any good reason why we should. So
that restriction was removed from both the fonts and images. CSS files
can reference fonts and images and we want those to work.

2. In development mode, we have a really sweet hack that gets rid of the
flash of unstyled content. You can see it in `getHead.js`. We get a list
of assets and do somethings… but webpack-isomorphic-tools replace
`node_modules` with `~` which was breaking that. So now it is swapped
back to `node_modules`. This is only development mode, so don't panic if
you don't like it but please share if there is a better way.