Add --debug, -D mode for server side debugging
Client side debugging is much easier because you can use the browser's
developer tools. In a universal app, sometimes you will need to debug
code that will be rendered on the server. `node-inspector` is a great
tool for letting you debug server side code in Blink web developer
tools. It's a little slow to start up but it can really save you a lot
of time debugging server side code.

[Node Inspector](https://github.com/node-inspector/node-inspector)