Expose specified environment variables to client
The application config file wasn't very useful because passing
environment variables into it made it so that it could only work on the
server side. If we wanted the environment variables to work on the
client side then webpack would need to know to expose them. **You
definitely don't want to just expose all environment variables or you
could leak things like api secrets**. It would also be annoying to have
to make an array of the variables that you wanted to expose.

To solve the problem, I decided to run static code analysis over the
config/application.js file to determine which environment variables the
developer was trying to use in the config/application.js file. I use
that to tell webpack which environment variables it needs to expose.

This was almost a complete solution until I realized that importing
config variables from this file in your application code would return
the entire object. So new projects will be generated with a slightly
modified application configuration file. I have also added an auto
upgrade script for backwards compatibility on all of the previously
generated applications.