Better support for api proxy
In the previous version of `gluestick-shared` when you made a request to
"/api" without a base URL, node had no idea what the actual URL to
request was. 0.3.9 resolves this by setting baseURL on http clients that
are set up with a request object.

We were also setting `changeOrigin:true` on the dev mode proxy. This was
making it so that the HTTP_HOST that the server would get would be
localhost:8880 instead of the url that you are hitting with. Now the
proper HTTP_HOST will be sent through