Improve existing action names
The previous action creator names for setting the server status codes
were sort of misleading with the name `render` in them. It made a user
think that some sort of React render was going on, when really we are
just setting the status code of the response on the server.