Stop using axios defaults for cookies
Axios was recycling the defaults object even in new axios instances so
it can not be trusted (for now).

We also had a fundamental issue with how we were treating incoming and
outgoing cookies. Cookeis that are part of a request object do not
include path, when we were sending the cookies with path, it was parsing
out the path as a separate cookie and mutating the desired cookies. Now
we treat incoming and outgoing cookies separately.