To make dnsmasq more portable
We should not override the container's resolv configuration (which might have
important configuration information in it). Instead, we add the nameserver
for localhost to be inserted above the container's name servers, which dnsmasq
will then use to resolve addresses.

Also, allow other files to be placed in /etc/dnsmasq.d so that subsequent
docker configuration layers can be added as needed.