force pull ofr node:6.10-slim (#973)
Our builds were using a cached version of node:6.10-slim which had an
outdated version of yarn in it. Even though we were trying to install
yarn the one installed via node:6.10-slim was favored so our version was
ignorred. If we pull the latest version it will maintain the yarn
updates for us and therefore we don't need to install yarn manually
anymore.